Handle prefix removal when entire path is removed
The req_path action allows for removing the entire matched path,
which could lead to an empty "path" variable in Varnish. This
did not render the config invalid but caused a 503 to be thrown
when the rule was triggered. The config template now includes a
check for an empty path variable and handles the case correctly.