Add support for SNI on backends
Currently ORM does not work with backends which require SNI headers to be sent,
such as cloudflare, istio-gateway.

This commit adds a new parameter on the origin object which can be used to
specify which SNI header to send. If the scheme is https then check-sni and sni
haproxy server options are set.

If the parameter is not set the behaviour is unchanged.